v1.0/requirements.txt
# Supply Chain v1.0 — Python Dependencies

# Core
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

# LLM + AI
google-generativeai>=0.8.0
openai>=1.50.0
langsmith>=0.1.0

# Memory (pgvector)
pgvector>=0.3.0

# Async Task Queue
celery>=5.4.0
redis>=5.0.0

# MCP (for MCP server only)
mcp[cli]>=1.0.0

# Utilities
pydantic>=2.0.0
